+++ src/camotics/qt/ToolGraphicsItem.cpp
@@ -26,79 +26,23 @@ using namespace CAMotics;
 
 
 ToolGraphicsItem::ToolGraphicsItem(QGraphicsItem *parent) :
-  QGraphicsPixmapItem(parent), number(0), highlight(false) {
-  setAcceptHoverEvents(true);
-  setCursor(Qt::OpenHandCursor);
-  setAcceptedMouseButtons(Qt::LeftButton);
+  QGraphicsPixmapItem(parent), number(0) {
 }
 
 
-void ToolGraphicsItem::update(const Tool &tool, int width, int height) {
+void ToolGraphicsItem::update(const Tool &tool, const QSize &size) {
   number = tool.getNumber();
 
   // Update tool view
   toolView.setTool(tool);
-  toolView.resize(width, height);
+  toolView.resize(size.width(), size.height());
   toolView.draw();
 
   // Paint image
   int stride = toolView.getStride();
   unsigned char *data = toolView.getBuffer().get();
-  QImage image(data, width, height, stride, QImage::Format_ARGB32);
+  QImage image(data, size.width(), size.height(), stride,
+               QImage::Format_ARGB32);
 
   setPixmap(QPixmap::fromImage(image));
-}
-
-
-void ToolGraphicsItem::hoverEnterEvent(QGraphicsSceneHoverEvent *event) {
-  highlight = true;
-  QGraphicsPixmapItem::update();
-}
-
-
-void ToolGraphicsItem::hoverLeaveEvent(QGraphicsSceneHoverEvent *event) {
-  highlight = false;
-  QGraphicsPixmapItem::update();
-}
-
-
-void ToolGraphicsItem::mousePressEvent(QGraphicsSceneMouseEvent *event) {
-  setCursor(Qt::ClosedHandCursor);
-}
-
-
-void ToolGraphicsItem::mouseReleaseEvent(QGraphicsSceneMouseEvent *event) {
-  setCursor(Qt::OpenHandCursor);
-}
-
-
-void ToolGraphicsItem::mouseMoveEvent(QGraphicsSceneMouseEvent *event) {
-#if 0
-  if (QLineF(event->screenPos(), event->buttonDownScreenPos(Qt::LeftButton))
-      .length() < QApplication::startDragDistance()) {
-    return;
-  }
-
-  QDrag *drag = new QDrag(event->widget());
-  QMimeData *mime = new QMimeData;
-  mime->setData("application/x-camotics-tool", QByteArray());
-  drag->setMimeData(mime);
-  drag->setPixmap(pixmap());
-  drag->setHotSpot(QPoint(pixmap().width() / 2, pixmap().height() / 2));
-  drag->exec();
-
-  setCursor(Qt::OpenHandCursor);
-#endif
-}
-
-
-void ToolGraphicsItem::paint(QPainter *painter,
-                             const QStyleOptionGraphicsItem *option,
-                             QWidget *widget) {
-  QGraphicsPixmapItem::paint(painter, option, widget);
-
-  if (highlight) {
-    painter->setPen(Qt::black);
-    painter->drawRect(boundingRect());
-  }
 }