+++ CHANGELOG.md
@@ -3,11 +3,15 @@
 ## v1.0.4:
  - Clear surface on reload.
  - Cutting move with zero feed is now an error.
+ - No error on rapid move with zero feed.
  - Allow tool zero.
  - Import/export tool table.
  - Improved tool table editing.
  - Attempt to fix missing tool paths in Windows 8.1 with Intel HD4000.
  - Default tool table.
+ - Simplified user interface.
+ - Moved project units and resolution to settings dialog.
+ - Default units, metric or imperial.
 
 ## v1.0.3:
  - Handle non-latin characters in filenames.